chore: rename spider_application module
